More package granularity.  Marked integration tests with .tag
